+++ apps/client/src/app/components/admin-market-data/admin-market-data.module.ts
@@ -4,10 +4,12 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatSortModule } from '@angular/material/sort';
 import { MatTableModule } from '@angular/material/table';
+import { RouterModule } from '@angular/router';
 import { GfActivitiesFilterModule } from '@ghostfolio/ui/activities-filter/activities-filter.module';
 
 import { AdminMarketDataComponent } from './admin-market-data.component';
 import { GfAssetProfileDialogModule } from './asset-profile-dialog/asset-profile-dialog.module';
+import { GfCreateAssetProfileDialogModule } from './create-asset-profile-dialog/create-asset-profile-dialog.module';
 
 @NgModule({
   declarations: [AdminMarketDataComponent],
@@ -15,10 +17,12 @@ import { GfAssetProfileDialogModule } from './asset-profile-dialog/asset-profile
     CommonModule,
     GfActivitiesFilterModule,
     GfAssetProfileDialogModule,
+    GfCreateAssetProfileDialogModule,
     MatButtonModule,
     MatMenuModule,
     MatSortModule,
-    MatTableModule
+    MatTableModule,
+    RouterModule
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })